requirements.txt
flask==3.0.3
spotipy==2.25.0
yt-dlp==2025.2.19
waitress
gunicorn
python-dotenv

